move Clock,Aferer,Nower from services to utils:
For easier reuse in other packages, to prevent import cycles.
